src/test/java/com/acompletenoobsmoke/mockito/exception_handling/BookServiceTest.java
package com.acompletenoobsmoke.mockito.exception_handling;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.sql.SQLException;
import java.time.LocalDate;

import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.Mockito.doThrow;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class BookServiceTest {

    @Mock
    private BookRepository bookRepository;

    @InjectMocks
    private BookService bookService;

    @Test
    public void testTotalPriceMethod() throws SQLException {
        when(bookRepository.getAllBooks()).thenThrow(SQLException.class);
        assertThrows(DatabaseReadException.class, () -> bookService.getTotalPriceOfBooks());
    }

    @Test
    public void testAddBook() throws SQLException {
        Book newBook = new Book("1111", "The Great Gatsby", 900, LocalDate.now());
        doThrow(SQLException.class).when(bookRepository).save(newBook);
        assertThrows(DatabaseWriteException.class, () -> bookService.addBook(newBook));
    }
}
